chore: rename define to helpers
