+++ packages/maskbook/src/plugins/FileService/SNSAdaptor/index.tsx
@@ -3,7 +3,7 @@ import type { Plugin } from '@masknet/plugin-infra'
 import { truncate } from 'lodash-es'
 import { base } from '../base'
 import { META_KEY_1, META_KEY_2 } from '../constants'
-import { FileInfoMetadataReader } from '../define'
+import { FileInfoMetadataReader } from '../helpers'
 import type { FileInfo } from '../types'
 import FileServiceDialog from './MainDialog'
 import { Preview } from './Preview'
