+++ packages/maskbook/src/plugins/FileService/Worker/database.ts
@@ -1,7 +1,7 @@
 import { createPluginDatabase } from '../../../database/Plugin/wrap-plugin-database'
 import { asyncIteratorToArray } from '../../../utils/type-transform/asyncIteratorHelpers'
 import { base } from '../base'
-import { FileInfoV1ToV2 } from '../define'
+import { FileInfoV1ToV2 } from '../helpers'
 import type { FileInfo, FileInfoV1 } from '../types'
 
 type TaggedTypes = FileInfo | FileInfoV1
