+++ packages/maskbook/src/plugins/FileService/helpers.ts
